+++ Illuminator.py
@@ -17,18 +17,26 @@ def run_nmap_ftp(ip, foldername):
     system("nmap --script=ftp-anon --script-args ftp-anon.maxlist=-1 -p 21 \
     %s 2>&1 > %s" % (ip,path.join(foldername, "FTP-Anonymous-Scan")))
     print "\t\t[-] FTP Anonymous Login Attempt Complete.\n"
+
+    #Attempt FTP Anonymous Directory Enumeration
+    print "\t[-] Test 2: FTP Anonymous Directory Enumeration..\n"
+    system("nmap -n -p 21 --script=ftp-anon.nse \
+    %s 2>&1 > %s" % (ip,path.join(foldername, "FTP-Anonymous-Dir-Enum")))
+    print "\t\t[-] FTP Anonymous Directory Enumeration Complete."
+
     #Attempt FTP Bruteforce
-    print "\t[-] Test 2: FTP Brute Force Attempt..\n"
+    print "\t[-] Test 3: FTP Brute Force Attempt..\n"
     system("nmap --script ftp-brute -p 21 \
     %s 2>&1 > %s" % (ip,path.join(foldername, "FTP-Bruteforce-Scan")))
     print "\t\t[-] FTP Brute Force Attempt Complete."
 
+
 #Function - Run Nmap NSE script for SMB User Enumeration
 def run_nmap_smb_users(ip, foldername):
-    print "\n[+] SMB Port 445 Detected - Starting SMB Enum Users Scan..\n"
-    system('nmap --script smb-enum-users.nse -p 139,445 192.168.62.134 | grep -v "Account disabled" \
+    print "\n[+] SMB Ports Detected - Starting SMB Enum Users Scan..\n"
+    system('nmap --script=smb-enum-users.nse -p 139,445 %s | grep -v "Account disabled" \
     | grep "Full name" | cut -d " " -f 7,10,11 2>&1 > %s' % (ip,path.join(foldername, "Nmap-SMB-Users")))
-    print "\t[-] SMB Enum Users NSE Scan Complete."
+    print "\t[-] SMB Enum Users Scan Complete."
 
 
 #Function - Run Nikto
@@ -54,11 +62,18 @@ def run_fierce(ip, foldername):
     print "\t[-] DNS enumeration Complete."
 
 
+#Function - Run SNMPCheck for SNMP
+def run_snmpcheck(ip, foldername):
+    print "\n[+] SNMP Port 161 Detected - Starting SNMPCheck Scan\n"
+    system("snmpcheck-1.9.rb %s 2>&1 > %s" % (ip,path.join(foldername, "SNMPCheck")))
+    print "\t[-] SNMP enumeration Complete."
+
+
 #Function - Run Enum4Linux
 def run_enum4linux(ip, foldername):
-    print "\n[+] Starting Host enumeration using Enum4Linux\n"
+    print "\n[+] Starting SMB enumeration using Enum4Linux\n"
     system("enum4linux %s 2>&1 > %s" % (ip,path.join(foldername, "Enum4Linux")))
-    print "  [-] Host enumeration Complete."
+    print "  [-] SMB enumeration Complete."
 
 
 #Function - Run Nmap scan
@@ -110,14 +125,14 @@ def process_scan(nmap_report):
 #Defining global variables
 scanned_ports = []
 port_map = {80:run_nikto,25:run_mailscan,110:run_mailscan,995:run_mailscan,
-            53:run_fierce,139:run_nmap_smb_users,21:run_nmap_ftp}
+            53:run_fierce,139:run_nmap_smb_users,21:run_nmap_ftp,161:run_snmpcheck}
 
 
 #Program Banner
 if len(sys.argv) <= 1:
     print(
 """
-Illuminator - Shed light onto darkness\nVersion: 1.2\nAuthor: Sam Rassam
+Illuminator - Shed light onto darkness\nVersion: 1.3\nAuthor: Sam Rassam @CyDef_Unicorn
 
 Illuminator utilizes the following tools:
 
@@ -125,7 +140,7 @@ Illuminator utilizes the following tools:
  * Nikto
  * Fierce
  * NBTScan
- * Snmpwalk
+ * SNMPCheck
  * Enum4Linux
 
 Usage: Illuminator.py <IP Address>
@@ -152,7 +167,7 @@ foldername = "Illuminator-%s" % ip
 mkdir (foldername)
 
 
-report = nmap_scan(ip, "-sV -n")
+report = nmap_scan(ip, "-n -sV --min-rate 5000 --max-retries 2 -p-")
 if report:
     process_scan(report)
     for scanned_port in scanned_ports:
@@ -162,7 +177,7 @@ if report:
     #Run enum4linux seperately
     run_enum4linux(ip, foldername)
     
-    print "\033[1;32;40m\n[**] Reports Completed [**]\n"
+    print "\n[**] Reports Completed [**]\n"
 else:
     print("No results returned")